[YS-453] refactor: 디바이스 경로에 따른 적응형 처리 middleware util 함수 분리
