+++ src/middleware/handlers/joinHandler.ts
@@ -1,13 +1,13 @@
-import { NextRequest, NextResponse } from 'next/server';
+import { NextRequest } from 'next/server';
 import type { JWT } from 'next-auth/jwt';
 
-import { getDeviceType, goToHome } from '../utils';
+import { goToHome, hasDevicePath, rewriteToDevicePath } from '../utils';
+
+const BASE_PATH = 'join';
 
 export function joinHandler(request: NextRequest, token: JWT | null) {
   const url = request.nextUrl.clone();
   const { searchParams, pathname } = url;
-  const userAgent = request.headers.get('user-agent') || '';
-  const deviceType = getDeviceType(userAgent);
 
   const isJoinSuccessPage = searchParams.get('step') === 'success';
 
@@ -16,16 +16,8 @@ export function joinHandler(request: NextRequest, token: JWT | null) {
     return goToHome(request);
   }
 
-  // desktop, mobile 구분
-  if (!pathname.match(/\/(join)\/(desktop|mobile)(\/.*)?$/)) {
-    const segments = pathname.split('/').filter(Boolean);
-    const basePath = segments[0];
-    const restPath = segments.length > 1 ? `/${segments.slice(1).join('/')}` : '';
-
-    const newPathname = `/${basePath}/${deviceType}${restPath}`;
-    url.pathname = newPathname;
-
-    return NextResponse.rewrite(url);
+  if (!hasDevicePath(pathname, BASE_PATH)) {
+    return rewriteToDevicePath(request, BASE_PATH, pathname);
   }
 
   return null;
